component structure refactored
