+++ main-app/src/components/layout/ResponsiveContainer.styles.scss
@@ -1,0 +1,25 @@
+@import "../../sass/variables";
+@import "../../sass/mixins";
+
+.responsiveContainer {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    padding: 16vw 0;
+    margin: 0 auto;
+}
+
+@include mdBreakPoint {
+    .responsiveContainer {
+        display: flex;
+        flex-direction: row;
+    }
+}
+
+@include xlBreakPoint {
+    .responsiveContainer {
+        margin: 0 auto;
+    }
+}
