+++ main-app/src/components/product-card/product-card.component.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { lazy, Suspense } from 'react';
-import PopUpImage from '../pop-up-image/pop-up-image.component'
+import PopUpImage from '../ui/pop-up-image/pop-up-image.component'
 // import CarouselSlider from '../carousel-slider/carousel-slider.component';
 import './product.card.style.scss';
 const CarouselSlider = lazy(() => import('../carousel-slider/carousel-slider.component'));
