+++ main-app/src/components/ui/counter/counter.component.tsx
