+++ main-app/src/components/ui/counter/counter.style.scss
@@ -1,5 +1,5 @@
-@import '../../sass/variables.scss';
-@import '../../sass/mixins';
+@import '../../../sass/variables.scss';
+@import '../../../sass/mixins';
 
 .info-card {
     display: flex;
