+++ main-app/src/components/ui/form-text-area-field/form-text-area.styles.scss
@@ -1,16 +1,16 @@
-@import '../../sass/variables.scss';
+@import '../../../sass/variables.scss';
 
 $main-color: $color-primary-4;
 $sub-color: $color-primary-0;
 $backgroundLocalColor: white;
 
 @mixin shrinkLabel {
-    top: -14px;
+    top: -20px;
     font-size: 12px;
     color: $main-color;
 }
 
-.form-input-wrapper {
+.form-textarea-wrapper {
     position: relative;
     margin: 45px 0;
 
@@ -18,29 +18,25 @@ $backgroundLocalColor: white;
         background: none;
         background-color: $backgroundLocalColor;
         color: $sub-color;
-        font-size: calc(0.7rem + 1vw);
+        font-size: 18px;
         padding: 10px 10px 10px 5px;
         display: block;
         width: 100%;
         border: none;
         border-radius: 0;
-        border-bottom: 1px solid $sub-color;
+        border: 1px solid $sub-color;
         margin: 25px 0;
 
         &:focus {
             outline: none;
         }
 
-        &:focus~.form-input-label {
+        &:focus~.form-textarea-label {
             @include shrinkLabel();
         }
     }
 
-    input[type='password'] {
-        letter-spacing: 0.3em;
-    }
-
-    .form-input-label {
+    .form-textarea-label {
         color: $sub-color;
         font-size: 16px;
         font-weight: normal;
