+++ main-app/src/components/ui/pop-up-image/pop-up-image.styles.scss
